Clean up metastore_search module and prep for integration test
