+++ cypress/support/commands/drupal.js
@@ -1,7 +1,5 @@
-/**
- * Drupal Collection
- */
-Cypress.Commands.add("drupalLogin", (user, password) => {
+// login as a given drupal user
+Cypress.Commands.add('drupalLogin', (user, password) => {
     return cy.request({
         method: 'POST',
         url: '/user/login',
@@ -14,11 +12,13 @@ Cypress.Commands.add("drupalLogin", (user, password) => {
     });
 });
 
+// logout of drupal
 Cypress.Commands.add('drupalLogout', () => {
     return cy.request('/user/logout');
 });
 
-Cypress.Commands.add("drupalDrushCommand", (command) => {
+// Run the supplied drush command
+Cypress.Commands.add('drupalDrushCommand', (command) => {
     var cmd = Cypress.env('drupalDrushCmdLine');
 
     if (cmd == null) {
