+++ cypress/support/index.js
@@ -14,7 +14,7 @@
 // ***********************************************************
 
 // Import commands.js using ES2015 syntax:
-import './commands'
+import './commands/drupal'
 
 // Alternatively you can use CommonJS syntax:
 // require('./commands')
